pyproject.toml
[project]
name = "try-dspy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dspy>=2.6.16",
    "python-dotenv>=1.1.0",
]
